Add building mechanics
This update does not include anything special. Just a note about
a pathfinding issue for a diamond building pattern, and basic setup for
putting building on a timer and snap buildings onto the grid from
the mouse position. Also allows a list of buildings for a player to
choose from as well as key combos for each building.